Moves google auth routes to routes folder
