+++ lib/auth/index.js
@@ -43,40 +43,40 @@ const plugin = {
     server.auth.strategy('simple', 'basic', { validate });
 
     server.route([
-      {
-        method: ['GET', 'POST'],
-        path: '/auth/google',
-        options: {
-          auth: 'google',
-          async handler(req, h) {
-            if (!req.auth.isAuthenticated) {
-              return `Authentication failed due to: ${req.auth.error.message}`;
-            }
+      // {
+      //   method: ['GET', 'POST'],
+      //   path: '/auth/google',
+      //   options: {
+      //     auth: 'google',
+      //     async handler(req, h) {
+      //       if (!req.auth.isAuthenticated) {
+      //         return `Authentication failed due to: ${req.auth.error.message}`;
+      //       }
   
-            const email = req.auth.credentials.profile.email;
-            if (!validateEmail(email)) {
-              return Boom.unauthorized('Intelligent error message goes here');
-            }
+      //       const email = req.auth.credentials.profile.email;
+      //       if (!validateEmail(email)) {
+      //         return Boom.unauthorized('Intelligent error message goes here');
+      //       }
   
-            req.cookieAuth.set({
-              token: req.auth.credentials.token,
-              profile: req.auth.credentials.profile,
-            });
-
-            return h.redirect('/restricted');
-          },
-        },
-      },
-      {
-        method: 'GET',
-        path: '/logout',
-        options: {
-          handler(req, h) {
-            req.cookieAuth.clear();
-            return h.redirect('/');
-          },
-        },
-      },
+      //       req.cookieAuth.set({
+      //         token: req.auth.credentials.token,
+      //         profile: req.auth.credentials.profile,
+      //       });
+      //       console.log('req.cookieAuth:', req.cookieAuth);
+      //       return h.redirect('/restricted');
+      //     },
+      //   },
+      // },
+      // {
+      //   method: 'GET',
+      //   path: '/logout',
+      //   options: {
+      //     handler(req, h) {
+      //       req.cookieAuth.clear();
+      //       return h.redirect('/');
+      //     },
+      //   },
+      // },
       {
         method: 'GET',
         path: '/auth/simple',
